[Sandbox] Move sandbox code to crate
